# Compute value hashes in parallel for v2Merkle