@dartdart26 2020-06-19
Closing this PR as per: https://github.com/vmware/concord-bft/pull/670#discussion_r442877274